+++ src/_internals/prop/index.js
@@ -1,0 +1,4 @@
+import _path from '../path';
+import curry from '../../curry';
+
+export default curry((p, obj) => _path([p], obj));