+++ src/pluck/index.js
@@ -1,31 +1,7 @@
+import _prop from '../_internals/prop';
 import curry from '../curry';
-import first from '../first';
-import is from '../is';
-
-// This will need to be updated to support list of object types
-const pluckObj = (a, list) =>
-  list.reduce((acc, v) => {
-    if (v[a]) {
-      acc.push(v[a]);
-    }
-
-    return acc;
-  }, []);
+import map from '../map';
 
 // Can be changed to better support array list types
 // Don't forget the documentation block
-export default curry((a, list) => {
-  if (Array.isArray(first(list))) {
-    return list.reduce((acc, v) => {
-      acc.push(v[a]);
-
-      return acc;
-    }, []);
-  }
-
-  if (is(Object, first(list))) {
-    return pluckObj(a, list);
-  }
-
-  return false;
-});
+export default curry((p, list) => map(_prop(p), list));